requirements.txt
opencv-python>=4.5.0
Pillow>=9.0.0
numpy>=1.21.0
matplotlib>=3.5.0
scipy>=1.7.0
scikit-image>=0.19.0
imageio>=2.19.0
